m365-file-organizer/requirements.txt
# Microsoft Graph
msgraph-sdk==1.5.4
azure-identity==1.15.0

# MongoDB
pymongo==4.6.1
motor==3.3.2

# AI
anthropic==0.25.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.9
httpx==0.26.0
tenacity==8.2.3
loguru==0.7.2

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
